+++ lib/export_instance.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:pdf/widgets.dart' as pw;
 
-import 'traversal.dart';
+import 'export_delegate.dart';
 import 'utils.dart';
 import 'widgets/container.dart';
 import 'widgets/center.dart';
