+++ lib/flutter_to_pdf.dart
@@ -6,5 +6,5 @@ export 'options/export_options.dart';
 export 'options/page_format_options.dart';
 export 'options/text_field_options.dart';
 export 'options/checkbox_options.dart';
-export 'traversal.dart';
+export 'export_delegate.dart';
 export 'export_frame.dart';
